+++ docs/modules/documentation/containers/button/examples/button-examples.component.ts
@@ -2,30 +2,35 @@ import { Component } from '@angular/core';
 
 @Component({
     selector: 'fd-button-options-example',
-    templateUrl: './button-options-example.component.html'
+    templateUrl: './button-options-example.component.html',
+    styleUrls: ['./button-examples.scss']
 })
 export class ButtonOptionsExampleComponent {}
 
 @Component({
     selector: 'fd-button-types-example',
-    templateUrl: './button-types-example.component.html'
+    templateUrl: './button-types-example.component.html',
+    styleUrls: ['./button-examples.scss']
 })
 export class ButtonTypesExampleComponent {}
 
 @Component({
     selector: 'fd-button-sizes-example',
-    templateUrl: './button-sizes-example.component.html'
+    templateUrl: './button-sizes-example.component.html',
+    styleUrls: ['./button-examples.scss']
 })
 export class ButtonSizesExampleComponent {}
 
 @Component({
     selector: 'fd-button-icons-example',
-    templateUrl: './button-icons-example.component.html'
+    templateUrl: './button-icons-example.component.html',
+    styleUrls: ['./button-examples.scss']
 })
 export class ButtonIconsExampleComponent {}
 
 @Component({
     selector: 'fd-button-state-example',
-    templateUrl: './button-state-example.component.html'
+    templateUrl: './button-state-example.component.html',
+    styleUrls: ['./button-examples.scss']
 })
 export class ButtonStateExampleComponent {}