+++ library/src/lib/panel/panel-description.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-    selector: 'fd-panel-description',
-    templateUrl: './panel-description.component.html'
-})
-export class PanelDescriptionComponent {}